\name{confintQRMissingBiBayesMix}
\alias{confintQRMissingBiBayesMix}
\title{Confint function}
\usage{
confintQRMissingBiBayesMix(mod)
}
\arguments{
  \item{mod}{}

  \item{...}{}
}
\value{
confint
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Minzhao Liu
}

