\name{ll2}
\alias{ll2}
\title{observednegative log likelihood}
\usage{
  ll2(param, y, X, R, tau, sp)
}
\arguments{
  \item{param}{}

  \item{y}{}

  \item{X}{}

  \item{R}{}

  \item{tau}{}

  \item{sp}{}
}
\value{
  negative log likelihood
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Minzhao Liu
}

