\name{ll2Mix}
\alias{ll2Mix}
\title{observed negative log likelihood}
\usage{
ll2Mix(param, y, X, R, tau, sp, K)
}
\arguments{
  \item{param}{}

  \item{y}{}

  \item{X}{}

  \item{R}{}

  \item{tau}{}

  \item{sp}{}
}
\value{
negative log likelihood
}
\description{
Observed Negative Log Likelihood
}
\details{
Give the observed nagetive log likelihood
}
\author{
Minzhao Liu
}

