\name{LLBiMix}
\alias{LLBiMix}
\title{observed log likelihood for bivariate case with mixture normal}
\usage{
LLBiMix(gamma1, beta1, gamma2, beta2sp, mu1, sigma1, mu2, sigma2, omega11,
  omega10, omega21, omega20sp, betay, betaysp, p, tau, y, X, R, K, G1, G2,
  model)
}
\arguments{
  \item{gamma1}{}

  \item{beta1}{}

  \item{sigma1}{}

  \item{gamma2}{}

  \item{beta2sp}{}

  \item{sigma21}{}

  \item{sigma21sp}{}

  \item{betay}{}

  \item{betaysp}{}

  \item{p}{}

  \item{tau}{}

  \item{y}{}

  \item{X}{}

  \item{R}{}
}
\value{
log likelihood of bivariate response with missingness
}
\description{
Observed Log Likelihood
}
\author{
Minzhao Liu, Mike Daniels
}

