\name{LLBiSingle}
\alias{LLBiSingle}
\title{observed log likelihood for bivariate case with single normal}
\usage{
  LLBiSingle(gamma1, beta1, sigma1, gamma2, beta2sp,
    sigma21, sigma21sp, betay, betaysp, p, tau, y, X, R)
}
\arguments{
  \item{gamma1}{}

  \item{beta1}{}

  \item{sigma1}{}

  \item{gamma2}{}

  \item{beta2sp}{}

  \item{sigma21}{}

  \item{sigma21sp}{}

  \item{betay}{}

  \item{betaysp}{}

  \item{p}{}

  \item{tau}{}

  \item{y}{}

  \item{X}{}

  \item{R}{}
}
\value{
  log likelihood of bivariate response with missingness
}
\description{
  Observed Log Likelihood
}
\author{
  Minzhao Liu, Mike Daniels
}

