\name{LLUniMix}
\alias{LLUniMix}
\title{observed log likelihood for univariate case with mixture of normals}
\usage{
LLUniMix(gamma, beta, K, mu, sigma, omega1, omega0, p, tau, y, X, R, G)
}
\arguments{
  \item{gamma}{}

  \item{beta}{}

  \item{p}{}

  \item{sigma}{}

  \item{tau}{}

  \item{y}{}

  \item{X}{}

  \item{R}{}
}
\value{
log likelihood of univariate response with missingness
}
\description{
Observed Log Likelihood
}
\author{
Minzhao Liu, Mike Daniels
}

