\name{LLUniSingle}
\alias{LLUniSingle}
\title{observed log likelihood for univariate case with single normal}
\usage{
  LLUniSingle(gamma, beta, p, sigma, tau, y, X, R)
}
\arguments{
  \item{gamma}{}

  \item{beta}{}

  \item{p}{}

  \item{sigma}{}

  \item{tau}{}

  \item{y}{}

  \item{X}{}

  \item{R}{}
}
\value{
  log likelihood of univariate response with missingness
}
\description{
  Observed Log Likelihood
}
\author{
  Minzhao Liu, Mike Daniels
}

